added function for asking players in player vs player for their name and clean up the code
I added a function where the program asks players in player vs player mode for their name. And also checks if the users' input are the same then asks the user again. And if the input is valid then it puts to an array. There are two arrays that I have set up. One for element and one for name. Which we can use when calling users or assigning elements to each players.