added a new component 8 to which is the end game part
In this change I have added a new component which is responsible for the end game mechanics, this component or function that I have added, checked if the user has won the round or if it was a tie. I also have added, a new function, where it asks the user if they want to play the game again. I still have not added, the computer's move, and also not added this to the main component as their is still going to be some changes that is going to be made in this component.