Added Component 9 (End Mechanics)
In this change I have made, I have added a new component which is  the component 9; End Mechanics. In this component, When the game/round is finished, the program will ask the user if they want to play the same they have played again, or play a different mode, or quit.