Rebuilt the base component program.
In this change, I have deleted all the work I have done so far to make the real base component of my game. I have thought about this in a while, since I have made several changes on my planning, and also to my functions. That is why I want to make a new start.

------------------------------------------

I have added an instruction.py, in this python file,  it consists the lists of the instructions that I will be called in one of my functions that I will be using in my base component