Created  component 11 (Text decorations) and added it tobase component
In this change I have made, I have added a new component which is for the text decorations . What I mean by the text decorations is the colours of the text, since I cannot change the size of the text.

The colours that I have added are red,  blue,  purple, darkcyan, green, yellow, and bold (white).